% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ak135_alp.R
\docType{data}
\name{ak135_alp}
\alias{ak135_alp}
\title{The ak135_alp velocity model}
\format{A TauP.R compatible velocity model}
\usage{
ak135_alp
}
\description{
The ak135_alp velocity model
}
\keyword{datasets}
