% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ak135_taupo.R
\docType{data}
\name{ak135_taupo}
\alias{ak135_taupo}
\title{The ak135_taupo velocity model}
\format{A TauP.R compatible velocity model}
\usage{
ak135_taupo
}
\description{
The ak135_taupo velocity model
}
\keyword{datasets}
