% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_all6station.R
\name{all6_station}
\alias{all6_station}
\title{Plot all6}
\usage{
all6_station(path, filter = c(1, 2, 3), zerophase = TRUE, E = ".e",
  N = ".n", Z = ".z")
}
\arguments{
\item{path}{Path to folder containing raw events and output folder}

\item{filter}{Which filter to plot (e.g. c(1,2,3) for fb1, fb2 & fb3)}

\item{zerophase}{Where the filters applied zero phase?}

\item{E}{Vector signal of the east component}

\item{N}{Vector signal of the north component}

\item{Z}{Vector signal of the vertical component}
}
\description{
A wrapper for the all6plot function to plot all events from a station
}
