% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_plotall6.R
\name{all6plot}
\alias{all6plot}
\title{Create all6 plot}
\usage{
all6plot(path, cuspid, filter = 1, zerophase = FALSE, E = ".e",
  N = ".n", Z = ".z", auto = FALSE)
}
\arguments{
\item{path}{Path to folder containing raw events and output folder}

\item{filter}{Which filter to plot (e.g. 1 for fb1)}

\item{zerophase}{Where the filters applied zero phase?}

\item{E}{Vector signal of the east component}

\item{N}{Vector signal of the north component}

\item{Z}{Vector signal of the vertical component}

\item{auto}{Select the first event if multiple are available?}

\item{display}{Display plots within R?}
}
\description{
Create an all6 plot for a specific event
}
