% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_anginc.R
\name{anginc}
\alias{anginc}
\title{Determines the angle of incidence for an event}
\usage{
anginc(tvel, trip)
}
\arguments{
\item{tvel}{Veloctity model read in by readtvel or a stored model (ak135_alp, ak135_taupo)}

\item{trip}{Seismogram triplet (output of readtriplet)}
}
\description{
Determines the angle of incidence for an event
}
