% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_butfilt.R
\name{butfilt}
\alias{butfilt}
\title{export}
\usage{
butfilt(a, fl = 0, fh = 0.5, deltat = 1, type = "BP", proto = "BU",
  npoles = 5, zp = TRUE, chebstop = 30, trbndw = 0.3, RM = FALSE)
}
\description{
export
}
