% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_checkcomponents.R
\name{checkcomp}
\alias{checkcomp}
\title{Check components}
\usage{
checkcomp(path, E = ".e", N = ".n", Z = ".z")
}
\arguments{
\item{path}{Path to folder}

\item{E}{Suffix of the east component}

\item{N}{Suffix of the north component}

\item{Z}{Suffix of the vertical component}
}
\description{
Checks a folder to make sure all three components are present and moves those with missing components to a subdirectory
}
