% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_checkspick.R
\name{checkspick}
\alias{checkspick}
\title{Check S-wave picks}
\usage{
checkspick(path, suffix = "E", header = "t0", E = ".e", N = ".n",
  Z = ".z")
}
\arguments{
\item{path}{Path to folder}

\item{suffix}{Which component to look for the S-pick in (E, N, or Z)}

\item{header}{Header name of where the S-pick is stored}

\item{E}{Suffix of the east component}

\item{N}{Suffix of the north component}

\item{Z}{Suffix of the vertical component}
}
\description{
Checks a folder to make sure all events have S-wave picks and moves those with missing picks to a subdirectory
}
