% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_create_ini.R
\name{createini}
\alias{createini}
\title{Create .ini}
\usage{
createini(path, trip, filts, name, number = 3, E = ".e", N = ".n",
  Z = ".z", nwbeg = 5, fdmin = 0.3, fdmax = 8, t_win_freq = 3,
  tlagmax = 1, Ncmin = 5, Mmax = 15, zerophase = FALSE)
}
\arguments{
\item{path}{Path to folder}

\item{trip}{Seismogram triplet (output of readtriplet)}

\item{name}{Event name (without suffix)}

\item{number}{Number of best filters to use}

\item{nwbeg}{number of start times tested}

\item{fdmin}{Minimum allowed dominant frequency}

\item{fdmax}{Maximum allowed dominant frequency}

\item{t_win_freq}{Window to calculate the dominant frequency}

\item{tlagmax}{Maximum allowed time delay (s)}

\item{Ncmin}{Minimum number of points in an acceptable cluster}

\item{Mmax}{maximum number ofclusters}

\item{A}{dataframe of the best filters (output of filter_spread)}

\item{suffe}{Suffix of east component}

\item{suffn}{Suffix of north component}

\item{suffz}{Suffix of vertical component}

\item{snrmax}{Minimum snr allowed for a good filter}

\item{t_win_snr}{Window for SNR (s)}

\item{t_err}{Modification to t_win_snr to account for error in S-pick (s)}
}
\value{
A vector of dominant frequency in the S-wave (maxfreq) for each filter
}
\description{
Creates an MFAST .ini (paramter) file
}
\examples{
# Create .ini file for event 2002.054.09.47.lhor2
pathto <- "~/mfast/sample_data/raw_data"
write_sample(pathto)
event <- "2002.054.09.47.lhor2"
triplet <- readtriplet(event,path=pathto)
bestfilt <- filter_spread(triplet)
maxfreq <- createini(pathto,triplet,bestfilt,event)
}
