% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_cutsimple.R
\name{cut_simple}
\alias{cut_simple}
\title{Simple cut}
\usage{
cut_simple(x, dt, t1, t2, b = 0)
}
\arguments{
\item{x}{vector signal}

\item{dt}{sample interval}

\item{t1}{Begin cut time}

\item{t2}{End cut time}
}
\value{
A cut vector signal
}
\description{
A simple routine to cuts out portion of a vector signal
}
