% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_type.R
\name{det.type}
\alias{det.type}
\title{Classify events}
\usage{
det.type(summ, cutoff = 15, cutoff2 = 35, minper = 0.9,
  depthcutoff = 0.17, sww = 45, tvel = ak135_taupo, plot = TRUE)
}
\arguments{
\item{summ}{A summary file (usually from a single station)}

\item{cutoff}{Maximum propgation angle allowed for a Type 1 event}

\item{cutoff2}{Maximum propgation angle allowed for a Type 2 event}

\item{minper}{Percentage of the path that must be below the maximum to be classified into that type}

\item{depthcutoff}{What percentage of the straight line path must be below cutoff2 to be classified as Type 2}

\item{sww}{The shear wave window cutoff}

\item{tvel}{The velocity model to be used (this must be the same as the one used to determine the ray parameters)}

\item{plot}{Should the result be plotted?}
}
\value{
A dataframe containing the summary file with the type for each event
}
\description{
Determines which type an event can be classified as
}
\details{
This function takes the events in a summary file and classifies each of them (based on their propogation angles) as Type 1 or Type 3 (full delay time and correct polarisation), Type 2 (correct polarisation) or Type NULL (low delay time and incorrect polarisation)
}
