% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_do_all_simple.R
\name{do_all_simple}
\alias{do_all_simple}
\title{Run MFAST on multiple folders}
\usage{
do_all_simple(path, sheader = "t0", type = "normal", filtnum = 3,
  tvelpath = NULL, tvel = ak135_alp, zerophase = TRUE,
  no_threads = NULL)
}
\arguments{
\item{path}{Path to folder containing folders with events}

\item{sheader}{SAC header the S-wave pick is stored in}

\item{type}{Which of the MFAST default settings and filters to use}

\item{filtnum}{Number of filters to test}

\item{tvelpath}{Path to a .tvel file containing the velocity model (overrides tvel)}

\item{tvel}{A tvel file read with readtvel (ak135_alp and ak135_taupo are already loaded)}

\item{no_threads}{Number of threads to run measurements on. Set to 1 for verbose mode. Defaults to automatic selection}
}
\value{
A dataframe containing a summary of all the stations
}
\description{
Run shear wave splitting measurements on more than one folder/station
}
\details{
Component suffixes are determined automatically
}
\examples{
# Run on measurements three folders of the normal sample data
write_sample("~/mfast/sample_data/raw_data")
write_sample("~/mfast/sample_data/raw_data2")
write_sample("~/mfast/sample_data/raw_data3")
do_all_simple(path="~/mfast/sample_data")
}
