% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_station_complex.R
\name{do_station_complex}
\alias{do_station_complex}
\title{Run shear wave splitting measurements on a folder of events with more options}
\usage{

  do_station_complex(path = "/home/stefan/mfast_package_v2.2/sample_data/raw_data",
  sheader = "t0", nwbeg = 5, fdmin = 0.3, fdmax = 8, t_win_freq = 3,
  tlagmax = 1, Ncmin = 5, Mmax = 15, snrmax = 3, t_win_snr = 3,
  t_err = 0.02, filtnum = 3, type = "normal", filter = NULL,
  tvelpath = NULL, tvel = ak135_alp, suffe = ".e", suffn = ".n",
  suffz = ".z")
}
\arguments{
\item{path}{Path to folder}

\item{sheader}{SAC header the S-wave pick is stored in}

\item{nwbeg}{number of start times tested}

\item{fdmin}{Minimum allowed dominant frequency}

\item{fdmax}{Maximum allowed dominant frequency}

\item{t_win_freq}{Window to calculate the dominant frequency}

\item{tlagmax}{Maximum allowed time delay}

\item{Ncmin}{Minimum number of points in an acceptable cluster}

\item{Mmax}{maximum number ofclusters}

\item{snrmax}{Minimum snr allowed to be processed}

\item{t_win_snr}{window for SNR}

\item{t_err}{Modification to t_win_snr to account for error in S-pick}

\item{filtnum}{Number of filters to test}

\item{type}{Which of the MFAST default settings and filters to use. If the P-wave pick is present, type="verylocal" uses the P-pick to set t_win_snr}

\item{filter}{User defined set of filters (this overrides the filter selected with type).}

\item{tvelpath}{Path to a .tvel file containing the velocity model (overrides tvel)}

\item{tvel}{A tvel file read with readtvel()}

\item{suffe}{Suffix of east component}

\item{suffn}{Suffix of north component}

\item{suffz}{Suffix of vertical component}
}
\description{
Run shear wave splitting measurements on a folder of events with more options
}
\examples{
# Run on measurements the normal sample data with defaults
write_sample("~/mfast/sample_data/raw_data")
do_station_complex(path="~/mfast/sample_data/raw_data")
}
