% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_weighteddt.R
\name{dt.weighted}
\alias{dt.weighted}
\title{Mean delay time}
\usage{
dt.weighted(summ, weights = c(1, 2, 3))
}
\arguments{
\item{summ}{Dataframe containing Castelazzi graded events (CZ_*.summ)}

\item{weights}{A vector containing the weights with length equal to the number of filters used (usually 3) in order with the first corresponding to F1}
}
\value{
A list containing the weighted mean delay time, and mean delay time per kilometre (straightline) path length as well as their respective standard deviations and standard errors.
}
\description{
Determine the mean weighted delay time
}
