% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_weightedfast.R
\name{fast.weighted}
\alias{fast.weighted}
\title{Mean fast polarisation}
\usage{
fast.weighted(summ, weights = c(1, 2, 3))
}
\arguments{
\item{summ}{Dataframe containing Castelazzi graded events (CZ_*.summ)}

\item{weights}{A vector containing the weights with length equal to the number of filters used (usually 3) in order with the first corresponding to F1}
}
\value{
A list containing the weighted mean polarisation, its pythagorean length, and the (weighted) p-value from the Rayleigh test
}
\description{
Determine the mean weighted fast polarisation
}
