% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_filter.R
\name{filter_spread}
\alias{filter_spread}
\title{Find best filters}
\usage{
filter_spread(trip, type = "normal", filter = NULL, t_win_snr = 3,
  t_err = 0.05, snrmax = 3, zerophase = TRUE)
}
\arguments{
\item{trip}{Seismogram triplet (output of readtriplet)}

\item{type}{Which of the default filters to use. If a P-wave pick is present, type="verylocal" uses it to set t_win_snr}

\item{filter}{User defined filters. Overrides filters selected by type (for "verylocal" the P-pick is still used)}

\item{t_win_snr}{Window for SNR}

\item{t_err}{Modification to t_win_snr to account for error in S-pick}

\item{snrmax}{Minimum snr allowed for a good filter}
}
\value{
A dataframe of the filters sorted by SNR*bandwidth
}
\description{
Determines the best filters for an event
}
\examples{
# Define your own set of filters
filt_low <- c(0.1,0.2,0.5)
filt_high <- c(1,2,3)
filts <- cbind(filt_low,filt_high)
write_sample("~/mfast/sample_data/raw_data")
triplet <- readtriplet("2002.054.09.47.lhor2",path="~/mfast/sample_data/raw_data")
bestfilt <- filter_spread(triplet,filter=filts)
}
