% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_getev.R
\name{getevents}
\alias{getevents}
\title{Get events}
\usage{
getevents(summ, events, station = NULL)
}
\arguments{
\item{summ}{Dataframe containing the summary file}

\item{events}{A vector containing the rquired event names}

\item{station}{Defaults to events on all stations}
}
\description{
A handy function to retrieve specific events from a summary dataframe
}
