% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_grade.R
\name{grade}
\alias{grade}
\title{Grade .summ file}
\usage{
grade(path, minsnr = 3, tlagmax = 1, minl = 0)
}
\arguments{
\item{path}{Path to .summ file to be graded}

\item{minsnr}{Minimum SNR allowed for an AB+ grade}

\item{tlagmax}{Maximum time delay allowed for an AB+ grade}

\item{minl}{Minimum lambdamax allowed for a AB+ grade}
}
\description{
Grades a .summ file (do_station automatically grades)
}
\examples{
# (Re)grade LHOR2.75.summ
write_sample("~/mfast/sample_data/raw_data")
do_station_simple(path="~/mfast/sample_data/raw_data")
pathto <- "~/mfast/sample_data/raw_data/LHOR2.summ_files/LHOR2.75.summ"
grade(pathto)
}
