% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_logfiles.R
\name{logfiles}
\alias{logfiles}
\title{Parse results}
\usage{
logfiles(path, name, trip, filtlist, maxfreqv, comment = "MFASTR", anginc)
}
\arguments{
\item{path}{Path to folder}

\item{name}{Name of event}

\item{trip}{Seismogram triplet (output of readtriplet)}

\item{filtlist}{Dataframe of the best filters to be used (output of writesac_filt)}

\item{maxfreqv}{Vector of dominant frequency in the S-wave (maxfreq) for each filter(output of create_ini)}

\item{comment}{Optional comment}

\item{anginc}{Angle of indidence (output of anginc)}
}
\value{
A dataframe containing the results for that event
}
\description{
Parses output of shear wave splitting measurement for a set of filters (used to build .summ files)
}
\examples{
# Run shear wave splitting measurement on event 2002.054.09.47.lhor2 and parse the results
pathto <- "~/mfast/sample_data/raw_data"
write_sample(pathto)
event <- "2002.054.09.47.lhor2"
triplet <- readtriplet(event,path=pathto)
a <- anginc(ak135_alp,triplet)
bestfilt <- filter_spread(triplet)
maxfreq <- createini(pathto,triplet,bestfilt,event)
f <- writesac_filt(pathto,triplet,event,bestfilt)
run_mfast(pathto,event,f)
res <- logfiles(pathto,event,triplet,f,maxfreq,anginc=a)
}
