% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_mclapply2.R
\name{mclapply2}
\alias{mclapply2}
\title{Wrapper around mclapply to track progress}
\usage{
mclapply2(X, FUN, ..., mc.preschedule = TRUE, mc.set.seed = TRUE,
  mc.silent = FALSE, mc.cores = getOption("mc.cores", 2L),
  mc.cleanup = TRUE, mc.allow.recursive = TRUE, mc.progress = TRUE,
  mc.style = 3)
}
\arguments{
\item{X}{a vector (atomic or list) or an expressions vector. Other
objects (including classed objects) will be coerced by
‘as.list’}

\item{FUN}{the function to be applied to}

\item{...}{optional arguments to ‘FUN’}

\item{mc.preschedule}{see mclapply}

\item{mc.set.seed}{see mclapply}

\item{mc.silent}{see mclapply}

\item{mc.cores}{see mclapply}

\item{mc.cleanup}{see mclapply}

\item{mc.allow.recursive}{see mclapply}

\item{mc.progress}{track progress?}

\item{mc.style}{style of progress bar (see txtProgressBar)}
}
\description{
Based on http://stackoverflow.com/questions/10984556
}
\examples{
x <- mclapply2(1:1000, function(i, y) Sys.sleep(0.01))
x <- mclapply2(1:3, function(i, y) Sys.sleep(1), mc.cores=1)
}
