% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_axialmean.R
\name{meanaxial}
\alias{meanaxial}
\title{Weighted axial mean}
\usage{
meanaxial(vec, weights = NULL)
}
\arguments{
\item{vec}{A vector of axis (degrees)}

\item{weights}{A vector of weights of the same length as vec}
}
\value{
The mean axis (degrees) and the Pythagorean length
}
\description{
The mean of a weighted axial variable
}
