% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_moving_dt.R
\name{moving_dt}
\alias{moving_dt}
\title{Delay time moving average}
\usage{
moving_dt(summfile, windowlength, windowspeed, norm = FALSE)
}
\arguments{
\item{summfile}{A dataframe containing a summary file (i.e. from readmfast)}

\item{windowlength}{Size of the averaging window (in days)}

\item{windowspeed}{Speed of advancing window (days per sample)}

\item{norm}{Normalise by straight line path distance?}
}
\value{
A dataframe containing the end days of each window along with its mean, standard deviation (of the mean), median, upper and lower 95% confidence intervals of the median, and the number of samples
}
\description{
A moving average of delay time
}
