% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_moving_vpvs.R
\name{moving_vpvs}
\alias{moving_vpvs}
\title{vP/vS moving average}
\usage{
moving_vpvs(vpvs, year, doy_det, windowlength, windowspeed)
}
\arguments{
\item{vpvs}{A vector of vP/vS ratios}

\item{year}{A vector of years for each vP/vS}

\item{doy_det}{A vector of julian days for each vP/vS}

\item{windowlength}{Size of the averaging window (in days)}

\item{windowspeed}{Speed of advancing window (days per sample)}
}
\value{
A dataframe containing the end days of each window along with its mean, standard deviation (of the mean), median, upper and lower 95% confidence intervals of the median, and the number of samples
}
\description{
A moving average of vP/Vs
}
