% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_parallel_simple.R
\name{parallel_simple}
\alias{parallel_simple}
\title{Parallel}
\usage{
parallel_simple(path, cores = NULL, sheader = "t0", type = "normal",
  filtnum = 3, tvelpath = NULL, tvel = ak135_alp)
}
\arguments{
\item{path}{Path to folder containing the folders for each station}

\item{cores}{Number of cores to use. Defaults to one less than the total}

\item{sheader}{SAC header the S-wave pick is stored in}

\item{type}{Which of the MFAST default settings and filters to use}

\item{filtnum}{Number of filters to test}

\item{tvelpath}{Path to a .tvel file containing the velocity model (overrides tvel)}

\item{tvel}{A tvel file read with readtvel (ak135_alp and ak135_taupo are already loaded)}
}
\value{
A list containing the summary files for all stations
}
\description{
Run shear wave splitting measurements on multiple stations in parallel
}
\details{
Component suffixes are determined automatically
}
\examples{
# Run on measurements the normal sample data in parallel
write_sample("~/mfast/sample_data/raw_data1")
write_sample("~/mfast/sample_data/raw_data2")
parallel_simple(path="~/mfast/sample_data")
}
