% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_pathclus.R
\name{pathclus}
\alias{pathclus}
\title{Path cluster}
\usage{
pathclus(summ, savepath, hvec = NULL, kmax = 7, runs = 20,
  minsample = 55, seed = NULL, plot = TRUE, rot = 180)
}
\arguments{
\item{summ}{An MFASTR summary file}

\item{savepath}{Path to save plots and files}

\item{hvec}{A vector of station elevations}

\item{kmax}{Maximum number of clusters}

\item{runs}{Number of runs for the clustering}

\item{seed}{Random number seed}

\item{plot}{Create plots?}

\item{rot}{Degrees to rotate 3D lower hemisphere plot}

\item{minsamples}{Minimum number of measurements for that station}
}
\value{
Creates folders containing the cuspids of events in each cluster along with the p-value of the Rayleigh test for polarisations in that cluster.
}
\description{
Clusters measurements by their station to event paths
}
\details{
Uses the movMF package to fit mixtures of von Mises Fisher distributions to the station to event paths projected onto a unit hemisphere below each station.
}
\examples{
# Run for all stations and save to clustest folder
cz <- summ.cz("~/summfiles")
pathclus(cz,savepath="~/clustest",plot=TRUE)
}
