% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_percentageanisotropy.R
\name{perani}
\alias{perani}
\title{Weighted percentage anisotropy}
\usage{
perani(summ, weights = NULL)
}
\arguments{
\item{summ}{Dataframe containing MFASt summary file}

\item{weights}{A vector containing the desired weights}
}
\value{
A dataframe containing each station and their corresponding percentage anistropy and shear wave anisotropy
}
\description{
Determine the weighted percentage anisotropy and shear wave anisotropy for each stations in a summary file
}
