% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_readMFAST.R
\name{readmfast}
\alias{readmfast}
\title{Read MFAST .summ file}
\usage{
readmfast(path, recuspid = FALSE)
}
\arguments{
\item{path}{The path the the summary file}

\item{recuspid}{Regenerate unique cuspids from event names? (Useful if they have been truncated in MFAST)}
}
\value{
A dataframe containing the summary file
}
\description{
Reads a .summ output from the original MFAST codes
}
\details{
This function is used with grade() to grade .summ files produced using the original MFAST codes (by setting mfast=TRUE).
}
