% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_readsac.R
\name{readtriplet}
\alias{readtriplet}
\title{Read a SAC format siesmogram triplet}
\usage{
readtriplet(event, path = ".", E = ".e", N = ".n", Z = ".z",
  header = "t0", pheader = "a", downsample = FALSE)
}
\arguments{
\item{event}{Event name}

\item{path}{Path to folder}

\item{E}{Suffix of east component}

\item{N}{Suffix of north component}

\item{Z}{Suffix of vertical component}

\item{header}{Name of header containing the S-wave pick}

\item{pheader}{Name of header containing the P-wave pick}

\item{downsample}{Downsample if sampling rate is less than 0.01s (Defaults to FALSE, originally used to decrease computational loads)}
}
\value{
A list containing dataframes for each of the three components with signal and header information
}
\description{
Reads, cuts, and loads S-wave pick into the t5 header using RSEIS/JSAC.seis as a workhorse
}
\details{
The S-wave pick must be stored on at least the east component and the P-wave pick (if present) must be stored on the vertical component
}
\examples{
# Read in 2002.054.09.47.lhor2
pathto <- "~/mfast/sample_data/raw_data"
write_sample(pathto)
event <- "2002.054.09.47.lhor2"
triplet <- readtriplet(event,path=pathto)
}
