% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_readtvel.R
\name{readtvel}
\alias{readtvel}
\title{Read .tvel}
\usage{
readtvel(name)
}
\arguments{
\item{name}{Name and path of .tvel file}
}
\value{
RSEIS compatible dataframe containing the velocity model
}
\description{
Reads a .tvel file and saves it in an RSEIS compatible format
}
\examples{
path <- "~/mfast/velocity/ak135_taupo.tvel"
model <- readtvel(path)
write_sample("~/mfast/sample_data/raw_data")
do_station_simple(path="~/mfast/sample_data/raw_data",tvel=model)
}
