% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_reanginc.R
\name{reanginc}
\alias{reanginc}
\title{Redetermine incidence angles}
\usage{
reanginc(summpath, tvel = ak135_taupo, overwrite = FALSE, mfast = FALSE,
  mc.cores = NULL)
}
\arguments{
\item{summpath}{Path to the .summ file}

\item{tvel}{Veloctity model read in by readtvel or a stored model (ak135_alp, ak135_taupo)}

\item{overwrite}{Should the original summfile be overwritten?}

\item{mfast}{Is the summfile from the original MFAST?}

\item{mc.cores}{Number of cores to run the calculations on (defaults to maximum)}
}
\value{
A summary file with redetermined incidence angles and ray parameters
}
\description{
Redetermine incidence angles for events in summary file
}
\examples{
# Redetermine the angle of incidences for a summary file
pathto <- "~/mfast/sample_data/summ_files/WPRZ.127.CZ.summ"
nsumm <- reanginc(pathto,tvel=ak135_alp)
}
