% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ_RK_NM.R
\docType{data}
\name{RK_NM_SWS}
\alias{RK_NM_SWS}
\title{RK_NM_SWS.summ}
\format{A dataframe containing MFAST the summary file for all Rotokawa and Ngatamariki template events}
\usage{
RK_NM_SWS
}
\description{
RK_NM_SWS.summ
}
\keyword{datasets}
