% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ_RK_NM_det.R
\docType{data}
\name{RK_NM_SWS_det}
\alias{RK_NM_SWS_det}
\title{RK_NM_SWS_det.summ}
\format{A dataframe containing the summary file for all Rotokawa and Ngatamariki matchfilter detected events}
\usage{
RK_NM_SWS_det
}
\description{
RK_NM_SWS_det.summ
}
\keyword{datasets}
