% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_mfast.R
\name{run_mfast}
\alias{run_mfast}
\title{Run splitting measurement}
\usage{
run_mfast(path, name, filtlist)
}
\arguments{
\item{path}{Path to folder}

\item{name}{Name of event}

\item{filtlist}{A dataframe of the best filters to be used (output of writesac_filt)}
}
\description{
Runs shearwave splitting measurements on a set of filtered SAC files
}
\examples{
# Run shear wave splitting measurements on event 2002.054.09.47.lhor2
pathto <- "~/mfast/sample_data/raw_data"
write_sample(pathto)
event <- "2002.054.09.47.lhor2"
triplet <- readtriplet(event,path=pathto)
bestfilt <- filter_spread(triplet)
maxfreq <- createini(pathto,triplet,bestfilt,event)
f <- writesac_filt(pathto,triplet,event,bestfilt)
run_mfast(pathto,event,f)
}
