% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_stde.R
\name{stde.weighted}
\alias{stde.weighted}
\title{Weighted standard error}
\usage{
stde.weighted(summ, weights = c(1, 2, 3), seed = NULL, iter = 9999)
}
\arguments{
\item{summ}{Dataframe containing Castelazzi graded events (CZ_*.summ)}

\item{weights}{A vector containing the weights with length equal to the number of filters used (usually 3) in order with the first corresponding to F1}

\item{seed}{A random number seed}

\item{iter}{Number of iterations}
}
\value{
The circular standard error in degrees
}
\description{
A bootstrapped weighted standard error for fast polarisations
}
