% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_readsumm.R
\name{summ.ab}
\alias{summ.ab}
\title{Read AB}
\usage{
summ.ab(path)
}
\arguments{
\item{path}{The path to the folder containing the .summ files}
}
\value{
A dataframe containing all the .summ files
}
\description{
Reads in multiple AB graded .summ files
}
