% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_readsumm.R
\name{summ.cz}
\alias{summ.cz}
\title{Read cz}
\usage{
summ.cz(path)
}
\arguments{
\item{path}{The path to the folder containing the .summ files}
}
\value{
A dataframe containing all the .summ files
}
\description{
Reads in multiple CZ graded .summ files
}
