% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_readsumm.R
\name{summ.null}
\alias{summ.null}
\title{Read null}
\usage{
summ.null(path)
}
\arguments{
\item{path}{The path to the folder containing the .summ files}
}
\value{
A dataframe containing all the .summ files
}
\description{
Reads in multiple null graded .summ files
}
