% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_writesample.R
\name{write_sample}
\alias{write_sample}
\title{Sample data}
\usage{
write_sample(path, type = "normal")
}
\arguments{
\item{path}{Path to folder}

\item{type}{"normal" or "verylocal" sample data}
}
\description{
Writes out MFAST sample data
}
\examples{
# Write out MFAST sample events
write_sample("~/mfast/sample_data/raw_data")
# Write out MFAST verylocal sample events
write_sample("~/mfast/sample_data/raw_data",type="verylocal")
}
