% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_writefiltsac.R
\name{writesac_filt}
\alias{writesac_filt}
\title{Write filtered SAC files}
\usage{
writesac_filt(path, trip, name, filtlist, number = 3, E = ".e", N = ".n",
  Z = ".z", zerophase = FALSE)
}
\arguments{
\item{path}{Path to folder}

\item{trip}{Event triplet (output of readtriplet)}

\item{name}{Name of the event}

\item{filtlist}{Dataframe of the best filters (output of filter_spread)}

\item{number}{Number of best filters to use}

\item{E}{Suffix of the east component}

\item{N}{Suffix of the north component}

\item{Z}{Suffix of the vertical component
#return A dataframe of the filters that have been written}
}
\description{
Writes out filtered waveforms ready to have shear wave splitting measured
}
\examples{
# Write out three best filters for event 2002.054.09.47.lhor2
pathto <- "~/mfast/sample_data/raw_data"
event <- "2002.054.09.47.lhor2"
write_sample(pathto)
triplet <- readtriplet(event)
bestfilt <- filter_spread(triplet)
f <- writesac_filt(pathto,triplet,event,bestfilt)
}
