% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_writefiltsac_simple.R
\name{writesac_filtsmp}
\alias{writesac_filtsmp}
\title{Simple write}
\usage{
writesac_filtsmp(path, trip, name, low, high, E = ".e", N = ".n",
  Z = ".z", n = 1)
}
\arguments{
\item{path}{Path to folder}

\item{trip}{Event triplet (output of readtriplet)}

\item{name}{Name of the event}

\item{low}{Low frequency cut-off}

\item{high}{High frequency cut-off}

\item{E}{Suffix of the east component}

\item{N}{Suffix of the north component}

\item{Z}{Suffix of the vertical component}

\item{n}{Number for suffix .fbn (e.g .fb2)}
}
\description{
Write out an event with a chosen filter
}
