% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_writetessa.R
\name{writetessa}
\alias{writetessa}
\title{Write TESSA .summ file}
\usage{
writetessa(summ, name)
}
\arguments{
\item{summ}{Dataframe containing the summary file of measurements to be run in TESSA}

\item{name}{Name of the file including path and .summ suffix (defaults to current working directory)}
}
\description{
Writes out a .summ file in the format required for TESSA
}
\examples{
# Create a .summ file for TESSA from all F1, F2 and F3 graded measurements
cz <- summ.cz("~/path/to/summfiles")
writetessa(cz,"~/TESSA/summfiles/cz.summ")
}
