% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_station_simple.R
\name{do_station_simple}
\alias{do_station_simple}
\title{Run MFAST}
\usage{
do_station_simple(path, sheader = "t0", type = "normal", filtnum = 3,
  tvelpath = NULL, tvel = ak135_alp)
}
\arguments{
\item{path}{Path to folder}

\item{sheader}{SAC header the S-wave pick is stored in}

\item{type}{Which of the MFAST default settings and filters to use}

\item{filtnum}{Number of filters to test}

\item{tvelpath}{Path to a .tvel file containing the velocity model (overrides tvel)}

\item{tvel}{A tvel file read with readtvel (ak135_alp and ak135_taupo are already loaded)}
}
\description{
Run shear wave splitting measurements on a folder of events
}
\details{
Component suffixes are determined automatically
}
\examples{
# Run on measurements the normal sample data
write_sample("~/mfast/sample_data/raw_data")
do_station_simple(path="~/mfast/sample_data/raw_data")

# Run on measurements the verylocal sample data where the S-pick is stored in the t5 header
write_sample("~/mfast/sample_data/raw_data",type="verylocal")
do_station_simple(path="~/mfast/sample_data/raw_data",type="verylocal",sheader="t5")
}
