% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_moving_dt.R
\name{moving_dt}
\alias{moving_dt}
\title{Delay time moving average}
\usage{
moving_dt(summfile, windowlength, windowspeed)
}
\arguments{
\item{summfile}{A dataframe containing a summary file (i.e. from readmfast)}

\item{windowlength}{Size of the averaging window (in days)}

\item{windowspeed}{Speed of advancing window (in days)}
}
\value{
A dataframe containing the end days of each window along with its mean and median delay time
}
\description{
A moving average of delay time
}
