% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_moving_phi.R
\name{moving_phi}
\alias{moving_phi}
\title{Fast polarisation moving average}
\usage{
moving_phi(summfile, windowlength, windowspeed)
}
\arguments{
\item{summfile}{A dataframe containing a summary file (i.e. from readmfast)}

\item{windowlength}{Size of the averaging window (in days)}

\item{windowspeed}{Speed of advancing window (days per sample)}
}
\value{
A dataframe containing the end days of each window along with its mean, median fast polarisation, and 95% confidence intervals of the mean (from a bootstrapped mle von Mises fit)
}
\description{
A moving average of fast polarisation
}
