% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_plotrose.R
\name{plotrose}
\alias{plotrose}
\title{Plot rose diagram}
\usage{
plotrose(path, summ, name = "rose.eps", bins = 16, kd = FALSE, sym = 16,
  prop = 1.3, bincol = "darkgrey", antibincol = "lightgrey",
  cols = "blue", antipodal = "lightblue", axes = TRUE, arrow = TRUE,
  arwcol = "red", arwlty = 1, arwlwd = 2)
}
\arguments{
\item{path}{Path to folder to save plots}

\item{name}{Name of plot}

\item{bins}{Number of bins}

\item{kd}{Kernal density?}

\item{sym}{Symbol for outer points}

\item{prop}{Scale length of rose diagram bins}

\item{bincol}{Colour of bins}

\item{cols}{Colour of points}

\item{antipodal}{Colour of antipodal points}

\item{axes}{Plot axes?}

\item{arrow}{Mean arrow? (Scaled by mean resultant length)}

\item{arwcol}{Arrow colour}

\item{arwlty}{Arrow line type (lty)}

\item{arwlwd}{Arrow line thickness (lwd)}

\item{bincol}{Colour of antipodal bins}
}
\description{
Plots a rose diagram of data from a .summ file
}
